requirements.txt
Flask==3.1.2
flask_cors==6.0.1
mongoengine==0.29.1
PyJWT==2.10.1
PyJWT==2.10.1
pymongo==4.15.0
python-dotenv==1.1.1
Requests==2.32.5
Werkzeug==3.1.3
